feat: centralize error handling and integrate transaction manager
