+++ crates/common/src/lib.rs
@@ -1,5 +1,7 @@
 pub mod constants;
+pub mod error;
 pub mod types;
 
 pub use constants::*;
+pub use error::*;
 pub use types::*;
