+++ crates/storage/Cargo.toml
@@ -6,7 +6,6 @@ edition = "2024"
 [dependencies]
 common.workspace = true
 crc32fast = "1.5.0"
-thiserror.workspace = true
 db-core.workspace = true
 [dev-dependencies]
 tempfile = "3.10.1"
