+++ crates/storage/src/buffer_pool/mod.rs
@@ -5,5 +5,6 @@ pub mod shard;
 #[cfg(test)]
 mod tests;
 
-pub use manager::{BufferPoolError, BufferPoolManager, Result};
+pub use common::BufferPoolError;
+pub use manager::{BufferPoolManager, Result};
 pub use shard::{BufferPoolShard, PageReadGuard, PageWriteGuard};
